% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{filter_by_prevalence}
\alias{filter_by_prevalence}
\title{Filter taxa by minimum non‐zero counts per group}
\usage{
filter_by_prevalence(Z, group, min_nonzero = 5)
}
\arguments{
\item{Z}{numeric matrix (samples x taxa)}

\item{group}{integer or factor vector of length nrow(Z), coded 0/1 (two groups)}

\item{min_nonzero}{integer; minimum required non‐zero samples per group}
}
\value{
filtered Z matrix
}
\description{
Filter taxa by minimum non‐zero counts per group
}
