% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gclr_test.R
\name{gclr.test}
\alias{gclr.test}
\title{gclr.test: Test arbitrary weighted taxa groups (GCLR inference)}
\usage{
gclr.test(model, variables, w1, w2)
}
\arguments{
\item{model}{The object fitted using function \code{McDisCov}.}

\item{variables}{Vector containing variables of interest.
For example: \code{variables = c("binDisease","cont")}.
For categorical variables, make sure the names align with the column names of \code{model$meta_data}.}

\item{w1}{Vector; weights of the taxa in the target taxa group. It will be scaled to unit sum automatically.}

\item{w2}{Vector; weights of the taxa in the reference taxa group. It will be scaled to unit sum automatically.}
}
\value{
A data frame with columns:
\itemize{
\item \code{covariate_name}: Variables of interest.
\item \code{estimate}: Effect size estimate in log scale for the target group relative to the reference group.
\item \code{SE}: Standard error of the estimate.
\item \code{TestStat}: Test statistic (chi-squared, 1 df).
\item \code{p_value}: p-value.
}
}
\description{
This function performs statistical hypothesis testing on taxa groups
(with arbitrary weighted components) for GCLR inference. It identifies taxa
groups that are significant in relative abundance (relative to the user-defined
reference taxa group).
}
\examples{
\dontrun{
# CLR
mcdiscov_clr <- function(model, variables) {
  res <- data.frame()
  for (i in seq(ncol(model$taxa_data))) {
    w1 <- rep(0, ncol(model$taxa_data))
    w2 <- rep(1, ncol(model$taxa_data))
    w1[i] <- 1
    res_i <- gclr.test(model, variables, w1 = w1, w2 = w2)
    res <- rbind(res, res_i)
  }
  rownames(res) <- colnames(model$taxa_data)
  return(res)
}

# ALR
mcdiscov_alr <- function(model, variables, reference = 1) {
  res <- data.frame()
  for (i in seq(ncol(model$taxa_data))) {
    w1 <- rep(0, ncol(model$taxa_data))
    w2 <- rep(0, ncol(model$taxa_data))
    w1[i] <- 1
    w2[reference] <- 1
    res_i <- gclr.test(model, variables, w1 = w1, w2 = w2)
    res <- rbind(res, res_i)
  }
  rownames(res) <- colnames(model$taxa_data)
  return(res)
}
}

}
\author{
Jiahao Wang (\email{10225000478@stu.ecnu.edu.cn})
}
