% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{gclr_transform}
\alias{gclr_transform}
\title{Generalized CLR transformation}
\usage{
gclr_transform(Z, selected_taxa, zero_impute = 0.5)
}
\arguments{
\item{Z}{count matrix (samples x taxa)}

\item{selected_taxa}{integer vector, indices of taxa for balance}

\item{zero_impute}{numeric, replacement for zeros}
}
\value{
log-ratio transformed matrix
}
\description{
Generalized CLR transformation
}
